+++ examples/data-sources/aiven_gcp_privatelink/data-source.tf
@@ -1,0 +1,4 @@
+data "aiven_gcp_privatelink" "main" {
+  project      = data.aiven_project.example_project.project
+  service_name = aiven_kafka.example_kafka.service_name
+}