+++ examples/resources/aiven_azure_privatelink_connection_approval/import.sh
@@ -1,1 +1,1 @@
-terraform import aiven_azure_privatelink_connection_approval.approval project/service_name
+terraform import aiven_azure_privatelink_connection_approval.approval PROJECT/SERVICE_NAME