+++ internal/sdkprovider/service/dragonfly/dragonfly.go
@@ -5,11 +5,10 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
 	"github.com/aiven/terraform-provider-aiven/internal/schemautil"
-	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig/dist"
 )
 
 func dragonflySchema() map[string]*schema.Schema {
-	s := schemautil.ServiceCommonSchema()
+	s := schemautil.ServiceCommonSchemaWithUserConfig(schemautil.ServiceTypeDragonfly)
 	s[schemautil.ServiceTypeDragonfly] = &schema.Schema{
 		Type:        schema.TypeList,
 		Computed:    true,
@@ -18,8 +17,6 @@ func dragonflySchema() map[string]*schema.Schema {
 			Schema: map[string]*schema.Schema{},
 		},
 	}
-	s[schemautil.ServiceTypeDragonfly+"_user_config"] = dist.ServiceTypeDragonfly()
-
 	return s
 }