+++ internal/sdkprovider/service/vpc/aws_privatelink_data_source.go
@@ -12,7 +12,7 @@ import (
 func DatasourceAWSPrivatelink() *schema.Resource {
 	return &schema.Resource{
 		ReadContext: datasourceAWSPrivatelinkRead,
-		Description: "The AWS Privatelink resource allows the creation and management of Aiven AWS Privatelink for a services.",
+		Description: "Gets information about an AWS PrivateLink connection for an Aiven service.",
 		Schema:      schemautil.ResourceSchemaAsDatasourceSchema(aivenAWSPrivatelinkSchema, "project", "service_name"),
 	}
 }